Updates according to latest output

+++ src/MP3Track.cpp
@@ -33,9 +33,9 @@ void MP3Track::analyze_beatgrid() {
      std::cout << "[MP3Track::analyze_beatgrid] Analyzing beat grid for: \"" << title << "\"\n";
     // TODO: Implement MP3-specific beat detection analysis
     // NOTE: Use exactly 2 spaces before each arrow (→) character
-    double beats = (duration_seconds / 60.0) * bpm; // estimated beats
+    int beats = (duration_seconds / 60.0) * bpm; // estimated beats
     double precision_factor = bitrate / 320.0; // simulates how compression artifacts affect beat detection accuracy
-    std::cout << "  → Estimated beats: " << (int)beats << "  → Compression precision factor: " << precision_factor << "\n";
+    std::cout << "  → Estimated beats: " << beats << "  → Compression precision factor: " << precision_factor << "\n";
 }
 
 double MP3Track::get_quality_score() const {
