+++ src/WAVTrack.cpp
@@ -29,8 +29,8 @@ void WAVTrack::analyze_beatgrid() {
     // 2. Calculate beats: (duration_seconds / 60.0) * bpm
     // 3. Print number of beats and mention uncompressed precision
     // should print "  → Estimated beats: <beats>  → Precision factor: 1.0 (uncompressed audio)"
-    double beats = (duration_seconds / 60.0) * bpm;
-    std::cout << "  → Estimated beats: " << (int)beats << "  → Precision factor: 1 (uncompressed audio)\n";
+    int beats = (duration_seconds / 60.0) * bpm;
+    std::cout << "  → Estimated beats: " << beats << "  → Precision factor: 1 (uncompressed audio)\n";
 }
 
 double WAVTrack::get_quality_score() const {
